Added _project_path built-in parameter for local-mode execution
