+++ digdag-core/src/main/java/io/digdag/core/agent/LocalAgentModule.java
@@ -17,7 +17,7 @@ public class LocalAgentModule
     public void configure(Binder binder)
     {
         binder.bind(AgentConfig.class).toProvider(AgentConfigProvider.class).in(Scopes.SINGLETON);
-        binder.bind(WorkspaceManager.class).to(NoopWorkspaceManager.class).in(Scopes.SINGLETON);
+        binder.bind(WorkspaceManager.class).to(LocalWorkspaceManager.class).in(Scopes.SINGLETON);
         binder.bind(TaskCallbackApi.class).to(InProcessTaskCallbackApi.class).in(Scopes.SINGLETON);
         binder.bind(TaskServerApi.class).to(InProcessTaskServerApi.class).in(Scopes.SINGLETON);
         binder.bind(OperatorManager.class).in(Scopes.SINGLETON);
